Sonar fixes

+++ Engine/Common/GeometryUtil.cs
@@ -1641,9 +1641,9 @@ namespace Engine.Common
 
                 var count = vertList.Count / sliceCount;
 
-                for (uint r = 0; r < count; r++)
+                for (uint r = 0; r < (uint)count; r++)
                 {
-                    for (uint i = 0; i < sliceCount; i++)
+                    for (uint i = 0; i < (uint)sliceCount; i++)
                     {
                         uint i0 = ((uint)sliceCount * r) + i;
                         uint i1 = ((uint)sliceCount * r) + ((i + 1) % (uint)sliceCount);
